# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MVP Kho Bac - Lenh Thanh Toan (LTT) API
  description: >
    REST API contract for the BFF-service (Backend-for-Frontend) managing
    Lenh Thanh Toan (Payment Orders) in the KBNN State Treasury system.

    Covers the full Maker-Checker-Approver lifecycle:
    Draft -> Ready_For_Approval -> Pending_Approver -> Approved -> Transferred_to_GL -> Posted

    Feature: FT-001 - Quan ly Lenh Thanh Toan
  version: 1.0.0
  contact:
    name: SA Agent - MVP Kho Bac
servers:
  - url: http://localhost:8080
    description: Local development BFF-service
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/ltt:
    get:
      operationId: listLtt
      summary: List and search Lenh Thanh Toan
      description: >
        Paginated listing of payment orders with multi-criteria filtering.
        Default sort: CREATED_DATE DESC. Default date range: last 7 days.
      tags: [LTT CRUD]
      parameters:
        - name: channel
          in: query
          description: "Payment channel: LIEN_NGAN_HANG or THANH_TOAN_SONG_PHUONG"
          schema:
            type: string
            enum: [LIEN_NGAN_HANG, THANH_TOAN_SONG_PHUONG]
        - name: transactionType
          in: query
          description: "Order type (dynamic based on channel)"
          schema:
            type: string
        - name: status
          in: query
          description: "Filter by status (repeat param for multi-select)"
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - DRAFT
                - READY_FOR_APPROVAL
                - PENDING_APPROVER
                - APPROVED
                - TRANSFERRED_TO_GL
                - POSTED
                - RETURNED_TO_MAKER
                - REJECTED
                - DELETED
        - name: refNo
          in: query
          description: "So YCTT / So but toan - exact or starts-with match"
          schema:
            type: string
        - name: senderCode
          in: query
          description: "NH/KB chuyen (sender bank/treasury code)"
          schema:
            type: string
        - name: receiverCode
          in: query
          description: "NH/KB nhan (receiver bank/treasury code)"
          schema:
            type: string
        - name: fromDate
          in: query
          description: "Start date for date range filter (inclusive)"
          schema:
            type: string
            format: date
            example: "2026-05-11"
        - name: toDate
          in: query
          description: "End date for date range filter (inclusive). Max range 90 days."
          schema:
            type: string
            format: date
            example: "2026-05-18"
        - name: dateField
          in: query
          description: "Which date column to apply the date range filter on"
          schema:
            type: string
            enum: [CREATED_DATE, PAYMENT_DATE, CHECKED_DATE, APPROVED_DATE]
            default: CREATED_DATE
        - name: amountFrom
          in: query
          description: "Minimum amount filter (inclusive)"
          schema:
            type: number
            format: double
            minimum: 0
        - name: amountTo
          in: query
          description: "Maximum amount filter (inclusive). Must be >= amountFrom."
          schema:
            type: number
            format: double
            minimum: 0
        - name: currencyCode
          in: query
          description: "Currency code (VND, USD, etc.)"
          schema:
            type: string
            default: VND
        - name: createdBy
          in: query
          description: "Username of the maker who created the LTT"
          schema:
            type: string
        - name: page
          in: query
          description: "Page number (zero-based)"
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: "Page size"
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: sort
          in: query
          description: "Sort criteria in format: field,direction (e.g. CREATED_DATE,DESC). Repeat for multi-sort."
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
            default: ["CREATED_DATE,DESC"]
      responses:
        "200":
          description: Paginated list of LTT summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttPageResponse"
              example:
                status: SUCCESS
                data:
                  content:
                    - fId: "LTT-2026-000001"
                      refNo: "YCTT-001234"
                      channel: LIEN_NGAN_HANG
                      transactionType: "LENH_THONG_THUONG"
                      lnhTransactionType: "GT_CAO_CO"
                      createdBy: "nguyenvt"
                      createdDate: "2026-05-18T09:30:00"
                      paymentDate: "2026-05-18"
                      senderCode: "01001"
                      senderName: "KBNN Ha Noi"
                      receiverCode: "02001"
                      receiverName: "KBNN TP HCM"
                      amount: 1500000000.00
                      currencyCode: VND
                      description: "Thanh toan hop dong so HD-2026-001"
                      fStatus: DRAFT
                      checkedBy: null
                      approvedBy: null
                      fVer: 1
                  totalElements: 1
                  totalPages: 1
                  page: 0
                  size: 20
                timestamp: "2026-05-18T09:35:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    # -------------------------------------------------------------------------
    # POST /api/v1/ltt - Create new LTT (Draft)
    # -------------------------------------------------------------------------
    post:
      operationId: createLtt
      summary: Create a new Lenh Thanh Toan (Draft)
      description: >
        Creates a new payment order in DRAFT status with F_VER=1.
        The system auto-assigns F_ID, CREATED_BY, CREATED_DATE.
        Supports idempotency via X-Request-ID header.
      tags: [LTT CRUD]
      parameters:
        - name: X-Request-ID
          in: header
          description: >
            Idempotency key. If a previous request with the same key succeeded,
            the system returns the previous response without re-processing.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLttRequest"
      responses:
        "201":
          description: LTT created successfully in DRAFT status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Duplicate X-Request-ID (idempotency conflict)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: ERROR
                code: "ERR-DUPLICATE-REQUEST"
                message: "A request with this X-Request-ID has already been processed."
                traceId: "abc123def456"
                timestamp: "2026-05-18T09:30:00"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # GET /api/v1/ltt/{id} - Get LTT detail
  # ---------------------------------------------------------------------------
  /api/v1/ltt/{id}:
    get:
      operationId: getLtt
      summary: Get LTT detail by F_ID
      description: >
        Returns the full payment order including header fields, detail lines (COA segments),
        sender info, receiver info, and attachment metadata.
      tags: [LTT CRUD]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
      responses:
        "200":
          description: Full LTT detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    # -------------------------------------------------------------------------
    # PUT /api/v1/ltt/{id} - Update LTT
    # -------------------------------------------------------------------------
    put:
      operationId: updateLtt
      summary: Update an existing LTT
      description: >
        Updates a payment order that is in DRAFT or RETURNED_TO_MAKER status.
        Only the original Maker may update. Uses optimistic locking via fVer field.
        On success, fVer increments by 1.
      tags: [LTT CRUD]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
        - name: X-Request-ID
          in: header
          description: Idempotency key for the update operation
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLttRequest"
      responses:
        "200":
          description: LTT updated successfully. fVer incremented.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Optimistic lock conflict (fVer mismatch) or LTT is not in an editable status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: ERROR
                code: "ERR-OPTIMISTIC-LOCK"
                message: "Ban ghi da bi thay doi tu phien khac. Vui long tai lai truoc khi tiep tuc."
                traceId: "abc123def456"
                timestamp: "2026-05-18T10:00:00"
        "500":
          $ref: "#/components/responses/InternalError"

    # -------------------------------------------------------------------------
    # DELETE /api/v1/ltt/{id} - Soft-delete LTT
    # -------------------------------------------------------------------------
    delete:
      operationId: deleteLtt
      summary: Soft-delete an LTT
      description: >
        Marks a payment order as DELETED (soft-delete). Only allowed for DRAFT or
        RETURNED_TO_MAKER status, and only by the original Maker.
        Requires a reason (min 10 chars) and confirmation that the record has been reviewed.
      tags: [LTT CRUD]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteLttRequest"
      responses:
        "200":
          description: LTT soft-deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                status: SUCCESS
                data:
                  fId: "LTT-2026-000001"
                  fStatus: DELETED
                  deletedBy: "nguyenvt"
                  deletedDate: "2026-05-18T10:30:00"
                timestamp: "2026-05-18T10:30:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: LTT is not in a deletable status (not DRAFT/RETURNED_TO_MAKER)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: ERROR
                code: "ERR-STATUS"
                message: "Giao dich dang o trang thai [READY_FOR_APPROVAL], khong cho phep Xoa."
                traceId: "abc123def456"
                timestamp: "2026-05-18T10:30:00"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # POST /api/v1/ltt/{id}/submit - Submit for approval
  # ---------------------------------------------------------------------------
  /api/v1/ltt/{id}/submit:
    post:
      operationId: submitLtt
      summary: Submit LTT for Checker review
      description: >
        Transitions LTT from DRAFT or RETURNED_TO_MAKER to READY_FOR_APPROVAL.
        Runs full validation before submission. Sends notification to Checker.
        Only the original Maker can submit.
      tags: [LTT Workflow]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
        - name: X-Request-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: LTT submitted successfully. Status is now READY_FOR_APPROVAL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: LTT not in submittable status or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # POST /api/v1/ltt/{id}/check - Checker approves/returns/rejects
  # ---------------------------------------------------------------------------
  /api/v1/ltt/{id}/check:
    post:
      operationId: checkLtt
      summary: Checker reviews LTT
      description: >
        Checker decision on a payment order in READY_FOR_APPROVAL status.
        - APPROVE: transitions to PENDING_APPROVER, notifies Approver.
        - RETURN: transitions to RETURNED_TO_MAKER, notifies Maker. Reason required (min 10 chars).
        - REJECT: transitions to REJECTED, locks the order. Reason required (min 10 chars).
        Enforces SoD: Checker must not be the same user as Maker.
      tags: [LTT Workflow]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
        - name: X-Request-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckLttRequest"
      responses:
        "200":
          description: Checker action processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: LTT not in READY_FOR_APPROVAL status, or SoD violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # POST /api/v1/ltt/{id}/approve - Approver final decision
  # ---------------------------------------------------------------------------
  /api/v1/ltt/{id}/approve:
    post:
      operationId: approveLtt
      summary: Approver final decision on LTT
      description: >
        Approver decision on a payment order in PENDING_APPROVER status.
        - APPROVE: transitions to APPROVED, triggers downstream workflow.
        - RETURN: transitions to RETURNED_TO_MAKER, notifies Maker. Reason required.
        - REJECT: transitions to REJECTED, locks the order. Reason required.
        Enforces SoD: Approver must not be the same as Maker or Checker.
        Requires authentication via OTP or digital signature.
      tags: [LTT Workflow]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
        - name: X-Request-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveLttRequest"
      responses:
        "200":
          description: Approver action processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: LTT not in PENDING_APPROVER status, SoD violation, or auth failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # POST /api/v1/ltt/{id}/copy - Copy LTT to new draft
  # ---------------------------------------------------------------------------
  /api/v1/ltt/{id}/copy:
    post:
      operationId: copyLtt
      summary: Copy an existing LTT to a new Draft
      description: >
        Creates a new payment order by copying data from an existing one.
        The new LTT gets a new F_ID, F_VER=1, F_STATUS=DRAFT.
        REF_NO must be updated by the Maker before submission.
        Certain fields (CREATED_BY, CREATED_DATE, audit info) are reset.
      tags: [LTT CRUD]
      parameters:
        - $ref: "#/components/parameters/LttIdPath"
        - name: X-Request-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: New LTT draft created from copy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LttDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # GET /api/v1/ltt/{id}/audit - Get audit history
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from the authentication endpoint.
        Token contains claims: sub (username), roles (MAKER/CHECKER/APPROVER), org (bank code).

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    LttIdPath:
      name: id
      in: path
      required: true
      description: "Unique identifier of the Lenh Thanh Toan (F_ID)"
      schema:
        type: string
        example: "LTT-2026-000001"

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request data or validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: ERROR
            code: "ERR-VALIDATION"
            message: "Vui long nhap [So YCTT]."
            traceId: "abc123def456"
            timestamp: "2026-05-18T09:30:00"
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: ERROR
            code: "ERR-AUTH"
            message: "Phien dang nhap da het han, vui long dang nhap lai."
            traceId: "abc123def456"
            timestamp: "2026-05-18T09:30:00"
    Forbidden:
      description: User lacks the required role or violates SoD
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: ERROR
            code: "ERR-PERMISSION"
            message: "Ban khong co quyen thuc hien thao tac nay."
            traceId: "abc123def456"
            timestamp: "2026-05-18T09:30:00"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: ERROR
            code: "ERR-SYSTEM"
            message: "Loi he thong, traceId: abc123def456. Vui long thu lai hoac lien he Quan tri."
            traceId: "abc123def456"
            timestamp: "2026-05-18T09:30:00"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope
      properties:
        status:
          type: string
          enum: [ERROR]
        code:
          type: string
          description: "Machine-readable error code"
          example: "ERR-VALIDATION"
        message:
          type: string
          description: "Human-readable error message in Vietnamese"
          example: "Vui long nhap [So YCTT]."
        traceId:
          type: string
          description: "Distributed tracing ID for log correlation"
          example: "abc123def456"
        timestamp:
          type: string
          format: date-time
          example: "2026-05-18T09:30:00"
        fieldErrors:
          type: array
          description: "Per-field validation errors (present when code=ERR-VALIDATION)"
          items:
            type: object
            properties:
              field:
                type: string
                example: "refNo"
              message:
                type: string
                example: "Vui long nhap So YCTT."
              rejectedValue:
                description: "The rejected value, if any"
      required: [status, code, message, traceId, timestamp]

    # =========================================================================
    # LTT Summary (used in list responses)
    # =========================================================================
...
    CreateLttRequest:
      type: object
      description: "Request to create a new payment order (Draft)"
      properties:
        channel:
          type: string
          enum: [LIEN_NGAN_HANG, THANH_TOAN_SONG_PHUONG]
          description: "Payment channel"
          example: "LIEN_NGAN_HANG"
        transactionType:
          type: string
          description: "Order type (depends on channel)"
          example: "LENH_THONG_THUONG"
        refNo:
          type: string
          description: "So YCTT / So but toan"
          maxLength: 50
          example: "YCTT-001234"
        paymentDate:
          type: string
          format: date
          description: "Ngay thanh toan (defaults to current date)"
          example: "2026-05-18"
        amount:
          type: number
          format: double
          description: "So tien chuyen (must equal sum of detail lines)"
          example: 1500000000.00
          minimum: 0
        currencyCode:
          type: string
          description: "Currency code"
          default: "VND"
          example: "VND"
        exchangeRate:
          type: number
          format: double
          description: "Ty gia (required only for foreign currency)"
          example: null
        lnhTransactionType:
          type: string
          description: "LNH transaction type (only when channel = LIEN_NGAN_HANG)"
          enum: [GT_CAO_CO, GT_CAO_NO, GT_THAP_CO, GT_THAP_NO]
          example: "GT_CAO_CO"
        orginNum:
          type: string
          description: "So chung tu goc (required when channel = TTSP)"
          example: null
        transactionDate:
          type: string
          format: date
          description: "Ngay chung tu (required when channel = TTSP)"
          example: null
        expType:
          type: string
          description: "Loai phi (required when transactionType = TT ngoai te khac)"
          example: null
        fnCode1:
          type: string
          description: "Ma ngoai te trich no"
          example: null
        fnCode2:
          type: string
          description: "Ma ngoai te TT"
          example: null
        fnAmount:
          type: number
          format: double
          description: "So tien ngoai te TT"
          example: null
        description:
          type: string
          description: "Noi dung thanh toan"
          example: "Thanh toan hop dong so HD-2026-001"
        details:
          type: array
          description: "Detail lines with COA segments"
          minItems: 1
          items:
            $ref: "#/components/schemas/LttDetailLineRequest"
        sender:
          $ref: "#/components/schemas/LttSenderRequest"
        receiver:
          $ref: "#/components/schemas/LttReceiverRequest"
        attachments:
          type: array
          description: "Attachment metadata (files uploaded separately via multipart endpoint)"
          items:
            type: object
            properties:
              docType:
                type: string
                enum: [CHUNG_TU_GOC, HOP_DONG, HOA_DON, BANG_KE, VAN_BAN_KHAC]
              note:
                type: string
                maxLength: 250
      required: [channel, transactionType, refNo, amount, currencyCode, description, details, sender, receiver]
...
    LttDetailResponse:
      allOf:
        - type: object
          properties:
            status:
              type: string
              enum: [SUCCESS]
            data:
              $ref: "#/components/schemas/LttDetail"
            timestamp:
              type: string
              format: date-time
          required: [status, data, timestamp]
    LttPageResponse:
      allOf:
        - type: object
          properties:
            status:
              type: string
              enum: [SUCCESS]
            data:
              type: object
              allOf:
                - type: object
                  properties:
                    content:
                      type: array
                      items:
                        $ref: "#/components/schemas/LttSummary"
                - $ref: "#/components/schemas/PageMetadata"
              required: [content]
            timestamp:
              type: string
              format: date-time
          required: [status, data, timestamp]
...